Group staves into systems
The algorithm is very simple. It relies on the fact that with the
-ddebug-skylines option, lilypond will issue skylines for systems:
a red one on top, and a green one below.

Hence, since one have the top and bottom limit of a system, one just
need to look if a staff fits inside.

A small error margin must be taken as it has been noticed that
sometimes, the system skylines don't cover the staff one by a very
tiny distance.