Grace notes are handled like normal notes. Avoid special cases.
This also fixes a problem when the start/stop time was not an integer.
This can happen when the key if e.g. 2/3. Guile would output the exact
value as a fraction (like 500/3) instead of a rounded value (like 167)

The grace note handling should probably be reworked as of right now,
they are placed such that they start at the exact same time as the
preceding note, making it (wrongly) sound like a chord.