Initialise bar numbers to max, the real value will be fetched from svg
In case of repeats appearing as volta repeats on the music sheets, the
bar numbers from the svg file won't match the ones from the notes file
as the latter gets unfolded repeats. One can think of it as if there
was two differents musics. One for the svg, one for the notes.