don't force the back-end to null
The back-end decides the format of the generated music sheet.
This can be one of pdf, svg, png ... Setting it to null disables the
generation of the music sheets, thus speeding up a lilypond pass.

This is particularly interesting when generating the notes files, as the
back-end (images) is then of no interested.

Unfortunately, forcing the back-end to null in the event listener also
prevents it to be overridden, even from command line.

The event-listener must be used when generating both the notes file, and
when generating the images. Therefore the back-end can't be set to null
there as it will prevent the creation of the images.

A solution would to set it to svg, and just discard the generated svg files
during the notes file generation pass. However as the svg files generation
takes a significant amount of time, the chosen solution is instead to
decide which back-end to use, from command-line, for both lilypond passes.