Generates boxes wrapping chords.
Each chord will now have a box that wrapps it. That box is delimited
by the left-most point of all notes in the chord, the right-most one,
and the top/bottom skyline of the system which contains the chord.

By using the top skyline and bottom skyline, one ensures that moving a
box showing which notes are currently played won't look weird. Indeed
it will show a cursor moving horizontally left to right. Using the top
and bottom-most notes in the chord would have the drawback that the
cursor would also move vertically, which would just hinder the user to
read the music sheet.