Fix a problem when a note is point is found to belong to two systems
The algorithm that detected the system to which a note belonged used
to look only at the y coordinate of the point and the top/bottom
skyline of the systems.
Looking at the max top/bottom has the advantage of quickly telling
apart if a point is in a system or not, but can actually return a
false positive.
Indeed, two systems can have their max_top/bottom skylines overlap
whereas actually the systems themselves, when looking at the full
top/bottom skylines don't.
Therefore when looking for a point in a system, if it appears that it
belongs to two or more systems (though it should never appear to
belong to more than two), the algorithm will now perform a more
accurate detection in order to discard the false positives.