Generate the keyboard events.
After reading the notes file, it now generates the key events.
Key events just say if a note is pressed or released.
This is going to be used by the player part.
The player part will read these and play the music.
The event also contains the staff number. It will determine the colour
for the keys. Two keys with the same staff number will have the same
colour. This way the keys meant to be played by the left hand will
be displayed in a different colour than the ones for the right hand.

So far, there is still something that isn't handled: non-piano voices.
For example, if a music sheet contains a piano staff (left and right
hands), but also a staff for a singer, the notes for the singer will
appear in the key events. That is because, for now, when the
event-listener generates the notes file, it doesn't output the
instrument name that must play it (piano, flute, voice ...)