Bar numbers are now extracted from the svg files
In the notes file, the repeats are unfolded. As a consequence, the bar
numbers won't match those in the svg files (for which the repeats are
kept untouched). The end-user will normally trust the bar numbers
displayed on the music sheet to be correct. Therefore the svg's bar
numbers "take precedence" over the one from the note file.

The event-listener will write the bar numbers at the beginning of the
grob id field, before "#origin=...". This way the program can extract
for each note the bar-number, the same way it already extracts the
x-width and y-height of the notes in the music sheet.