The svg file saved on the binary file doesn't have skylines any more.
To get svg files with skylines (for data extraction), and without ones
(for music sheet display), the generation of svg files will have one
more pass.

Previously it had one pass to get the notes file. On this pass all
repeats are unfolded.
It also used to have a pass which gets svg files with debug skylines.
Now, it will have a third pass to generate the svg without the debug
skylines.

Once the pass that does the no-skylines svg files is done, all files
it generated are renamed XXX.svg.no_skylines. This way, there is an
easy association between file with and without skylines: just append
".no_skylines" to the filename

The binary must be called only with the svg filenames that do contain
the skylines, and it will automatically pick up the ones without when
appropriate.