Bugfix: correct number of events at time
In the output file format, events happening at the same time are
grouped together. The output starts by the number of events happening
at the same time, and then the events are dumped.

The problem with the former calculation is that it would not take into
account the case of turn-page-events. These events are not passed as
arguments to the function, but are generated on the fly when dumping
the cursor events. Therefore, the function that computes the number of
events at time, must also recreate the logic to detect if a turn page
event will be generated or not.