Ensure the pointer is set to NULL on exit.
In case the scope_exit macro fails to create the object wrapper, the
code would have exited without setting the pointer back to null.
By creating the SCOPE_EXIT macro before changing the pointer, one
ensures that the object wrapper is set properly set when changing the
pointer, and thus, the pointer will be set back to null on exit.