Added code for the library to PRELOAD when executing lilypond.
This LD_PRELOAD library will hook on the open function, and when
lilypond opens the music-functions.scm, will patched the file on the
fly so that repeats are forced to be unfolded. This works with the
music-functions.scm file from lilypond 2.19. It doesn't work with
earlier versions of lilypond. This is due to the patching system
being very primitive and dumb at this point.