Revert "Factorised code and do not use the event-listener to get skylines"
This reverts commit 04f0b2803aa5774023a55253b37ac8376a9caf34.
The event-listener serves two purposes when generating the skylines:
  - setting the optionto request skylines
  - setting the id on the notes in the svg file

The former commit that removed the use of the event listener when
generating skylines took care of using the option to request them,
but forgot to set the id on the notes.
This commit fixes this by reverting to the old behaviour of using
the event-listener.
However, the removal of the event-listener for that specific use-case
was due to an issue. It will try to use the notes and the instruments
files with the default name. In some circumstances this would end up
in an error of the type "no-such-file" and crash.
To address this issue, this commit creates a new option to disable the
generation of the notes and table file.