Handle empty colored group generated by lilypond.
In some cases, lilypond generate a "g" node in the svg file, with a
color property that matches the ones for skylines. However the node
do not contain anything below. As a consequence, lilydumper would
detect an invalid line and throw.
This commit workaround the issue by skipping these empty nodes.