Create notes and sn2in file with perms from lilydumper
This is a workaround to avoid the event listener to do it itself.
It has been noted that the guile part can fail in some weird way when
creating these files. E.g. setting the permissions for the sn2in file
to 100 (--x------). Consequence being then the program can't read the
file back.