Handle group nodes with id but no child path generated by lilypond
In some cases lilypond generate a 'g' node containing an id but
without a path node in it. As a consequence, the former cursor box
extractor would sometime fails with error "a note with id ... could
not be found in an svg file" because it only looked for 'g' nodes
with a path inside.
The fix is simply to also allow nodes without path but to initialise
the bouding box in a way that as soon as there is another note in the
same chord, the cursor box will get the correct values from the
bounding boxes of the other notes.

Then add a sanity check to ensure the get_cursor_box never returns
a cursor box with invalid left/right/top/bottom coordinates