Bugfix: handle ids with a space in them
In some corner case, the duration string of a note is 'log = -1'.
Because it has space in it, the formated string extracting (using
stringstream) would therefore stop extracting the id at that point
(the space after 'log'). As a consequence the id would be incompletely
read, and later the program would fail as it would say it couldn't
find a note with such an id in the svg.

This commit fixes that by properly extracting the id. It is however
important to note that this implementation relies on the fact that
the id field is the last one on the line. Otherwise some stronger
parsing would be required.