Removed useless code as note heads can appear in only one svg file.
Before, the program would try to find all pages that include a note
head, and then find the common one with all pages containing the other
note heads appearing in the same chord. And do some sanity check then.
Now, since a note head should only appear in one svg file, it just
finds the page containing the first note head of the chord, and then
make sure all other notes of the chord are on the same page. As
a consequence it also provide a better error message