Handle a weird corner case.
In one case, lilypond outputted a different staff-number for a note
when generating the svg file than when generating the notes file.
Turned out the correct staff number was the one in the note file, so
as a consequence, this is the one that the program will keep, and the
staff-number won't be part of the id anymore.