Bugfix: stronger check for finding staves.
It is possible to have small horizontal lines in the middle of a staff.
Therefore, the precedent check, that was only looking for five lines
in a row would fail because not all of the five were staff lines.

Now the check is stronger as avoids these possible little lines to
prevent matching a staff.