add --include to directory of ly file when calling lilypond
This basicaly add supports for music sheets split in several files
(so using the \include) directive. Otherwise, it would have been
necessary to first change directory to the folder of the source file
before calling lilydumper. Failing to do so would have resulted in
lilypond complaining that it couldn't find the included files.