Bugfix: ensures grace notes won't exceed grace_note_length.
In case grace notes were between two normal notes, the former code
would just allocate the full time between the normal notes to grace
notes. As a consequence, if the two normal note before the grace notes
would be separated by a long time, e.g. if it was a white or round
note the grace notes would then get a long time, thus not really being
like a grace note.

This commit, thus allocate a maximum time for the grace notes, and
place them near the end.